feat: Access control for admin panel

+++ convex/betterAuth/api.ts
@@ -5,4 +5,4 @@ import schema from './schema';
 // Use factory function for options to satisfy type signature
 export const api = createApi(schema, () => getBaseAuthOptions());
 
-export const { create, findOne, findMany, updateOne, updateMany, deleteOne, deleteMany } = api;
+export const { create, findOne, findMany, updateOne, updateMany, deleteOne, deleteMany, getCorePermissions, getAdminDashboardStats, getCurrentUserWithPermissions, getAdminUserList, getAdminOrgList, getMembersWithUsers } = api;
